rag_chatbot/requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langgraph[sqlite]
gradio
chromadb
sentence-transformers
requests
python-dotenv